Rename the PingFlow class to be clearer
