+++ ping-pong/cordapp/workflows/src/main/kotlin/net/corda/examples/pingpong/flows/PingFlow.kt
@@ -1,4 +1,4 @@
-package net.corda.examples.pingpong
+package net.corda.examples.pingpong.flows
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.flows.*
